+++ demos/demo_operation_log.sh
@@ -7,7 +7,7 @@ new_tmp_dir
     jj git clone https://github.com/octocat/Hello-World
     cd Hello-World
     jj abandon octocat-patch-1
-    jj branch forget octocat-patch-1
+    jj branch forget --global octocat-patch-1
 } > /dev/null
 
 comment "We are in the octocat/Hello-World repo.